test: Modify to support trivy template option
